feat: 楽曲詳細ページからプレイリストに曲を追加、削除する機能を追加

+++ src/app/api/musicDeletePlaylist/route.ts
@@ -2,36 +2,33 @@ import prisma from "@/lib/prisma";
 import { type NextRequest, NextResponse } from "next/server";
 
 type Body = {
-  deleteSubmitPlaylists: { playlist_id: number }[];
+  diffPlaylists: { playlist_id: number; music_flag: boolean }[];
   music_id: number;
 };
 
 export const DELETE = async (req: NextRequest) => {
   try {
     const body: Body = await req.json();
-    const deletePlaylists = body.deleteSubmitPlaylists;
+    const diffPlaylists = body.diffPlaylists;
     const music_id = body.music_id;
     if (!body) {
       throw new Error("楽曲データの受け渡しに失敗しました");
     }
-
-    for (let i = 0; i < deletePlaylists.length; i++) {
-      const submitCheck = await prisma.playlist_Song.findFirst({
-        select: { id: true },
-        where: {
-          playlist_id: deletePlaylists[i].playlist_id,
-          api_song_id: BigInt(music_id),
-        },
-      });
-
-      if (submitCheck) {
-        await prisma.playlist_Song.delete({
+    for (let i = 0; i < diffPlaylists.length; i++) {
+      if (!diffPlaylists[i].music_flag) {
+        const submitCheck = await prisma.playlist_Song.findFirst({
           where: {
-            // playlist_id: deletePlaylists[i].playlist_id,
-            // api_song_id: BigInt(music_id),
-            id: submitCheck.id,
+            playlist_id: diffPlaylists[i].playlist_id,
+            api_song_id: BigInt(music_id),
           },
         });
+        if (submitCheck) {
+          await prisma.playlist_Song.delete({
+            where: {
+              id: submitCheck.id,
+            },
+          });
+        }
       }
     }
